refactor: Added a package for analysis.computational. Moved TreeVisitor and computationals into their designeted diractory.
