+++ src/main/java/de/featjar/feature/model/analysis/computation/ComputeUniformity.java
@@ -18,7 +18,7 @@
  *
  * See <https://github.com/FeatureIDE/FeatJAR-feature-model> for further information.
  */
-package de.featjar.feature.model.analysis;
+package de.featjar.feature.model.analysis.computation;
 
 import de.featjar.analysis.javasmt.computation.ComputeSatisfiability;
 import de.featjar.analysis.javasmt.computation.ComputeSolutionCount;
