+++ src/main/java/de/featjar/feature/model/cli/PrintStatistics.java
@@ -31,15 +31,15 @@ import de.featjar.feature.model.FeatureModel;
 import de.featjar.feature.model.IFeatureModel;
 import de.featjar.feature.model.IFeatureTree;
 import de.featjar.feature.model.analysis.AnalysisTree;
-import de.featjar.feature.model.analysis.ComputeFeatureAverageNumberOfChildren;
-import de.featjar.feature.model.analysis.ComputeFeatureFeaturesCounter;
-import de.featjar.feature.model.analysis.ComputeFeatureGroupDistribution;
-import de.featjar.feature.model.analysis.ComputeFeatureTopFeatures;
-import de.featjar.feature.model.analysis.ComputeFeatureTreeDepth;
-import de.featjar.feature.model.computation.ComputeAtomsCount;
-import de.featjar.feature.model.computation.ComputeAverageConstraint;
-import de.featjar.feature.model.computation.ComputeFeatureDensity;
-import de.featjar.feature.model.computation.ComputeOperatorDistribution;
+import de.featjar.feature.model.analysis.computation.ComputeAtomsCount;
+import de.featjar.feature.model.analysis.computation.ComputeAverageConstraint;
+import de.featjar.feature.model.analysis.computation.ComputeFeatureAverageNumberOfChildren;
+import de.featjar.feature.model.analysis.computation.ComputeFeatureDensity;
+import de.featjar.feature.model.analysis.computation.ComputeFeatureFeaturesCounter;
+import de.featjar.feature.model.analysis.computation.ComputeFeatureGroupDistribution;
+import de.featjar.feature.model.analysis.computation.ComputeFeatureTopFeatures;
+import de.featjar.feature.model.analysis.computation.ComputeFeatureTreeDepth;
+import de.featjar.feature.model.analysis.computation.ComputeOperatorDistribution;
 import de.featjar.feature.model.io.FeatureModelFormats;
 import de.featjar.feature.model.io.csv.CSVAnalysisFormat;
 import de.featjar.feature.model.io.json.JSONAnalysisFormat;
